requirements.txt
async-timeout==4.0.3
dnspython==2.6.1
Faker==25.2.0
mongoengine==0.28.2
pika==1.3.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
redis-lru==0.1.2
six==1.16.0
